Convert SelfAssessment/Info to TypeScript
Why:
----
We would like to have all of our files using typescript.

This Commit:
----
- Convert the SelfAssessment/Info file to TypeScript
- Rename `Info` to `AssessmentLayout` since it reflects more closely
what the file is doing

Reviewers:
----
This is a major effort around converting the files into TypeScript, this
is one commit of many, we are focusing first on having an integrated
type system and then we can tackle the behavior of these components.